refactor(bot): split queueManipulation.ts (961 LOC) into focused modules
Extract four focused modules from the 961-line hotspot:
- trackNormalization.ts: normalizeTrackKey, normalizeText, genre family utilities
- queueEditOps.ts: clearQueue, shuffleQueue, smartShuffleQueue, removeTrackFromQueue,
  moveTrackInQueue, extractSpotifyTrackId, markAsAutoplayTrack, moveUserTrackToPriority,
  blendAutoplayTracks
- candidateFallback.ts: collectBroadFallbackCandidates, collectGenreCandidates,
  enrichWithAudioFeatures, interleaveByArtist, CandidateContext
- queueRescue.ts: rescueQueue, buildVcContributionWeights, getHistoryTracks,
  QueueRescueResult, RescueQueueOptions

queueManipulation.ts becomes a thin barrel (~90 LOC) that re-exports everything
plus retains getTrackAudioFeatures. All existing exports preserved for backward
compatibility. 2865 tests pass, tsc --noEmit clean.
